Update and rename LegendaryBot.py to RegiBot.py
